backend/requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
requests==2.32.5
backboard-sdk==1.4.7
google-genai==1.57.0
openai==2.15.0
gunicorn==23.0.0
